+++ mocks/wasmd/x/wasm/keeper/query_plugins.go
@@ -1,7 +1,0 @@
-package keeper
-
-import "github.com/cosmos/cosmos-sdk/baseapp"
-
-type GRPCQueryRouter interface {
-	Route(path string) baseapp.GRPCQueryHandler
-}