+++ scripts/test_cover.sh
@@ -6,16 +6,11 @@ SUBMODULES=$(find . -type f -name 'go.mod' -print0 | xargs -0 -n1 dirname | sort
 CURDIR=$(pwd)
 echo "mode: atomic" > coverage.txt
 
-if [ "$EXPERIMENTAL" == "true" ]
-then
-    EXPERIMENTAL_TAG="experimental"
-fi
-
 for m in ${SUBMODULES[@]}; do
     cd $CURDIR/$m
     PKGS=$(go list ./...)
     for pkg in ${PKGS[@]}; do
-        go test -v -timeout 30m -coverpkg=all -coverprofile=profile.out -covermode=atomic -tags="norace ledger test_ledger_mock $EXPERIMENTAL_TAG" "$pkg"
+        go test -v -timeout 30m -coverpkg=all -coverprofile=profile.out -covermode=atomic -tags="norace ledger test_ledger_mock" "$pkg"
         if [ -f profile.out ]; then
             tail -n +2 profile.out >> $CURDIR/coverage.txt;
             rm profile.out